Memperbaiki tipe data pada migrasi 2020-07-17-174430 Table Address Customer

+++ database/migrations/2020_07_17_174430_create_table_address_customer.php
@@ -15,7 +15,7 @@ class CreateTableAddressCustomer extends Migration
     {
         Schema::create('address_customer', function (Blueprint $table) {
             // Create table address_customer
-            $table->integer('id_customer')->unsigned()->index();
+            $table->bigInteger('id_customer')->unsigned()->index();
             $table->integer('id_address')->unsigned()->index();
             $table->timestamps();
